# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Efficace API",
        "description": "API de la plateforme Efficace",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.100.102:8000",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/loginWorker": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d'un travailleur",
                "operationId": "4713b41c04e6c0924e55e7d39ad39b9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "worker@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "4|rRyQ76Jaqy0..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Lister les jobs du travailleur connecté",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "Liste des jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Construction Villa Tagidor"
                                                    },
                                                    "start_date": {
                                                        "type": "string",
                                                        "example": "2026-03-10"
                                                    },
                                                    "end_date": {
                                                        "type": "string",
                                                        "example": "2026-06-30"
                                                    },
                                                    "workers_found": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Voir les détails d'un job",
                "operationId": "6de31489b41fa51255446ac2ad15b7fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du job",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Construction Villa Tagidor"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Construction d'une villa R+2"
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "2026-03-10"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "example": "2026-06-30"
                                                },
                                                "localisation": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Douala"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "workers": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 28
                                                            },
                                                            "firstName": {
                                                                "type": "string",
                                                                "example": "Jean"
                                                            },
                                                            "lastName": {
                                                                "type": "string",
                                                                "example": "Dupont"
                                                            },
                                                            "worker": {
                                                                "type": "string",
                                                                "example": "architect"
                                                            },
                                                            "lot": {
                                                                "type": "string",
                                                                "example": "Gros oeuvre"
                                                            },
                                                            "note": {
                                                                "type": "integer",
                                                                "example": 8
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Supprimer un job",
                "operationId": "5537fd998abc3a62eef8c09e4ae52000",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}